+++ react/hooks/useSimulation.ts
@@ -45,7 +45,7 @@ type Params = {
   inView: boolean
   onComplete: (product: ProductSummaryTypes.Product) => void
   onError: () => void
-  priceBehavior: 'async' | 'default'
+  priceBehavior: 'async' | 'asyncOnly1P' | 'default'
 }
 
 function useSimulation({
@@ -57,22 +57,25 @@ function useSimulation({
 }: Params) {
   const items = product.items || []
 
-  const simulationItemsInput = useMemo(
-    () =>
-      items.map((item) => ({
+  const simulationItemsInput = useMemo(() => {
+    if (priceBehavior === 'async') {
+      return items.map((item) => ({
         itemId: item.itemId,
-        sellers: item.sellers.map((seller) => ({
-          sellerId: seller.sellerId,
-        })),
-      })),
-    [items]
-  )
+        sellers: item.sellers.map((seller) => ({ sellerId: seller.sellerId })),
+      }))
+    }
+
+    return items.map((item) => ({
+      itemId: item.itemId,
+      sellers: [{ sellerId: '1' }],
+    }))
+  }, [items, priceBehavior])
 
   useQuery(QueryItemsWithSimulation, {
     variables: {
       items: simulationItemsInput,
     },
-    skip: priceBehavior !== 'async' || !inView,
+    skip: priceBehavior === 'default' || !inView,
     ssr: false,
     onError,
     onCompleted: (response) => {
@@ -87,13 +90,41 @@ function useSimulation({
       mergedProduct.items.forEach((item, itemIndex) => {
         const simulationItem = simulationItems[itemIndex]
 
-        const sellerDefault = getDefaultSeller(simulationItem.sellers)
+        if (priceBehavior === 'async') {
+          const sellerDefault = getDefaultSeller(simulationItem.sellers)
 
-        item.sellers = item.sellers.map((seller, simulationIndex) => {
-          const sellerSimulation = simulationItem.sellers[simulationIndex]
+          item.sellers = item.sellers.map((seller, simulationIndex) => {
+            const sellerSimulation = simulationItem.sellers[simulationIndex]
 
-          return mergeSellers(seller, sellerSimulation, sellerDefault)
-        })
+            return mergeSellers(seller, sellerSimulation, sellerDefault)
+          })
+        } else {
+          const seller1PIndex = item.sellers.findIndex(
+            (seller) => seller.sellerId === '1'
+          )
+
+          const sellers = Array.from(item.sellers)
+
+          sellers[seller1PIndex] = simulationItem.sellers[0]
+          const sellerDefault = getDefaultSeller(sellers)
+
+          item.sellers = item.sellers.map((seller) => {
+            if (seller.sellerId !== '1') {
+              return !sellerDefault
+                ? seller
+                : {
+                    ...seller,
+                    sellerDefault: seller.sellerId === sellerDefault,
+                  }
+            }
+
+            return mergeSellers(
+              seller,
+              simulationItem.sellers[0],
+              sellerDefault
+            )
+          })
+        }
       })
 
       mergedProduct.sku = mergedProduct.items.find(
@@ -101,9 +132,9 @@ function useSimulation({
       ) as ProductSummaryTypes.SingleSKU
 
       if (mergedProduct.sku.sellers.length > 0) {
-        mergedProduct.sku.seller = mergedProduct.sku.sellers.find(
+        mergedProduct.sku.seller = (mergedProduct.sku.sellers.find(
           (seller) => seller.sellerDefault
-        ) as ProductSummaryTypes.Seller
+        ) ?? mergedProduct.sku.sellers[0]) as ProductSummaryTypes.Seller
       } else {
         mergedProduct.sku.seller = {
           // @ts-expect-error We are not providing the full type