+++ CHANGELOG.md
@@ -8,10 +8,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- I18n pseudo-language to implement In Context tool.
+- I18n Bg and Bs.
 
 ### Fixed
-- I18n Fr, IT and Ro.
+- I18n Fr, It, Ja, Ko and Ro.
+
+## [2.76.0] - 2021-09-17
+### Added
+- The ability to dynamically choose the SKU that will be initially selected
+
+## [2.75.0] - 2021-08-23
+
+### Added
+- `asyncOnly1P` to `priceBehavior`.
 
 ## [2.74.1] - 2021-08-09